Move `ember-views/system/utils` tests to new infrastructure.
